Changed the unordered list to order list.
I have changed the unordered list to ordered list as it was more appropriate that will help in figuring them out properly. As well as I have made a few more changes including bolding the phrase "a dwarf planet " to bring attention that Pluto now has been assigned as a dwarf planet and building " Pluto " in the tile as it will help in bringing attention to the main motive of this webpage that is it is about the dwarf planet-Pluto.